feat(platforms): adding more info to Holonym sidebar
